Update AnsiTerminal.ino
Alt+Ctrl+F12 wasnt working (for me) anymore
Please check if this is also Ok for you

+++ examples/VGA/AnsiTerminal/AnsiTerminal.ino
@@ -115,7 +115,8 @@ void setup()
       } else {
         // pressing CTRL + ALT + F12, reset parameters and reboot
         if ((Terminal.keyboard()->isVKDown(VirtualKey::VK_LCTRL) || Terminal.keyboard()->isVKDown(VirtualKey::VK_RCTRL)) &&
-            (Terminal.keyboard()->isVKDown(VirtualKey::VK_LALT) || Terminal.keyboard()->isVKDown(VirtualKey::VK_RALT))) {
+            (Terminal.keyboard()->isVKDown(VirtualKey::VK_LALT) || Terminal.keyboard()->isVKDown(VirtualKey::VK_RALT)) &&
+            (Terminal.keyboard()->isVKDown(VirtualKey::VK_F12)))  {
           Terminal.deactivate();
           preferences.clear();
           // show reboot dialog
